Introduce `InspectableGraphStore`.
